% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleR.R
\docType{data}
\name{cell.type.classification}
\alias{cell.type.classification}
\title{Reference matrix for Kang et al. classification}
\format{A list with the reference matrix}
\usage{
cell.type.classification
}
\description{
Reference matrix for Kang et al. classification
}
\keyword{datasets}
