% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{clusters.map.values}
\alias{clusters.map.values}
\title{Map clusters to values}
\usage{
clusters.map.values(cluster_ids, singler_clusters)
}
\arguments{
\item{cluster_ids}{the cluster ids of each single cell}

\item{singler_clusters}{the cluster annotation for each cluster}
}
\value{
annotation for each single cell
}
\description{
Map clusters to values
}
