% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleR.R
\name{Combine.Multiple.10X.Datasets}
\alias{Combine.Multiple.10X.Datasets}
\title{Helper function to combine multiple 10X datasets}
\usage{
Combine.Multiple.10X.Datasets(dirs, random.sample = 0, min.genes = 500)
}
\arguments{
\item{dirs}{a list of the directories of the 10X samples. dirs should be the absolute path. dirs = list.dirs(dir.path,full.names = T)}

\item{random.sample}{number of cells to choose randomly from each sample}

\item{min.genes}{a threshold on the number of non-zero genes.}
}
\value{
a list with sc.data - the count matrix, and orig.ident - the directory name of the sample
}
\description{
Helper function to combine multiple 10X datasets
}
