% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleR.Create.R
\name{CreateBigSingleRObject}
\alias{CreateBigSingleRObject}
\title{Analyze very big data sets. Runs SingleR on small chunks (10,000 cells per run) and then combines them together.}
\usage{
CreateBigSingleRObject(counts, annot = NULL, project.name, xy, clusters,
  N = 10000, min.genes = 200, technology = "10X",
  species = "Human", citation = "", ref.list = list(),
  normalize.gene.length = F, variable.genes = "de", fine.tune = T,
  reduce.file.size = T, do.signatures = F, do.main.types = T,
  temp.dir = getwd(), numCores = SingleR.numCores)
}
\arguments{
\item{counts}{a tab delimited text file containing the counts matrix, a 10X directory name or a matrix with the counts.}

\item{annot}{a tab delimited text file or a data.frame. Rownames correspond to column names in the counts data}

\item{project.name}{the project name}

\item{xy}{a matrix with the xy coordinates. From the original single-cell object.}

\item{clusters}{the clusters identities.From the original single-cell object.}

\item{N}{number of cells in each iteration. Default is 10000.}

\item{min.genes}{Include cells where at least this many genes are detected (number non-zero genes).}

\item{technology}{The technology used for creating the single-cell data.}

\item{species}{The species of the sample ('Human' or 'Mouse').}

\item{citation}{a citation for the project.}

\item{ref.list}{a list of reference objects. If NULL uses the predefined reference objects - Mouse: ImmGen and Mouse.RNAseq, Human: HPCA and Blueprint+Encode.}

\item{normalize.gene.length}{if a full-length method set to TRUE, if a 3' method set to FALSE.}

\item{variable.genes}{variable gene method to use - 'sd' or 'de'. Default is 'de'.}

\item{fine.tune}{perform fine tuning. Default is TRUE. Fine-tuning may take long to run.}

\item{do.signatures}{create signatures data}

\item{do.main.types}{run the SingleR pipeline for main cell types (cell types grouped together) as well.}

\item{temp.dir}{used by the SingleR webtool.}

\item{numCores}{Number of cores to use.}

\item{clusters}{input cluster id for each of the cells with at least min.genes, if NULL uses SingleR clusterings.}
}
\description{
Analyze very big data sets. Runs SingleR on small chunks (10,000 cells per run) and then combines them together.
}
