% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleR.R
\name{CreateVariableGeneSet}
\alias{CreateVariableGeneSet}
\title{Create a list for differential genes for all pairwise cell types in the a reference data set.
Using this significantly reduces computation time for the 'de' variable gene method.}
\usage{
CreateVariableGeneSet(ref_data, types, n)
}
\arguments{
\item{ref_data}{the reference dataset gene expression matrix}

\item{types}{labels for each column in ref_data}

\item{n}{number of genes per pairwise comparison}
}
\value{
list of lists. Each list contains a list for each cell types with n differential genes.
}
\description{
Create a list for differential genes for all pairwise cell types in the a reference data set.
Using this significantly reduces computation time for the 'de' variable gene method.
}
