% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{datasets}
\alias{datasets}
\alias{HumanPrimaryCellAtlasData}
\alias{BlueprintEncodeData}
\alias{ImmGenData}
\alias{MouseRNAseqData}
\alias{DatabaseImmuneCellExpressionData}
\alias{NovershternHematopoieticData}
\alias{MonacoImmuneData}
\title{Reference dataset extractors}
\usage{
HumanPrimaryCellAtlasData(...)

BlueprintEncodeData(...)

ImmGenData(...)

MouseRNAseqData(...)

DatabaseImmuneCellExpressionData(...)

NovershternHematopoieticData(...)

MonacoImmuneData(...)
}
\arguments{
\item{...}{Further arguments to pass to the \pkg{celldex} function of the same name.}
}
\value{
A \link[SummarizedExperiment]{SummarizedExperiment} object containing the reference dataset.
}
\description{
These dataset getter functions are deprecated as they have been migrated to the \pkg{celldex} package
for more general use throughout the Bioconductor package ecosystem.
}
\author{
Aaron Lun
}
