% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleR.R
\name{fineTuningRound}
\alias{fineTuningRound}
\title{Internal function for a step in the fine-tuning process}
\usage{
fineTuningRound(topLabels, types, ref_data, genes, mat, sd.thres, sc_data,
  quantile.use, fine.tune.thres)
}
\arguments{
\item{topLabels}{list of the cell type to use in the fine-tuning round}

\item{types}{list of all cell types}

\item{ref_data}{the reference expression data}

\item{genes}{list of genes to use for the analysis, or 'de' or 'sd'}

\item{mat}{the median expression matrix for the reference}

\item{sd.thres}{the standard deviation threshold to use for choosing genes. Only used if genes == 'sd'}

\item{sc_data}{the single cell data}

\item{quantile.use}{the quantile for the aggregating scores}

\item{fine.tune.thres}{a threshold of top cell types to use for the next round.}
}
\value{
the cell types to use for the next round or one cell type which is the final annotation.
}
\description{
Internal function for a step in the fine-tuning process
}
