% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleR.R
\docType{data}
\name{human.egc}
\alias{human.egc}
\alias{human.egc}
\title{Human signatures}
\format{A GeneSetCollection of 5 signatures}
\usage{
human.egc

human.egc
}
\description{
Human signatures

Human signatures
}
\keyword{datasets}
