% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleR.R
\docType{data}
\name{human_lengths}
\alias{human_lengths}
\title{Length of human genes for TPM calculation}
\format{A named vector of gene lengths}
\usage{
human_lengths
}
\description{
Length of human genes for TPM calculation
}
\keyword{datasets}
