% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleR.R
\name{medianMatrix}
\alias{medianMatrix}
\title{Calculate median values by groups}
\usage{
medianMatrix(mat, groups)
}
\arguments{
\item{mat}{a matrix of numeric values NxM}

\item{groups}{a vector of size M with annotation of columns in mat}
}
\value{
a matrix with one column per group
}
\description{
Calculate median values by groups
}
