% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{MonacoImmuneData}
\alias{MonacoImmuneData}
\title{Obtain human bulk RNA-seq data of sorted cell populations from GSE107011}
\usage{
MonacoImmuneData()
}
\value{
A \linkS4class{SummarizedExperiment} object with a \code{"logcounts"} assay
containing the log-normalized expression values, along with cell type labels in the 
\code{\link{colData}}.
}
\description{
Download and cache the normalized expression values of 114 bulk RNA-seq samples
of sorted immune cell populations that can be found in 
\href{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE107011}{GSE107011}.
}
\details{
The dataset entails 114 human RNA-seq samples annotated to 11 main cell types: 
\itemize{
    \item CD8+ T cells
    \item T cells
    \item CD4+ T cells
    \item Progenitors
    \item B cells
    \item Plasmablasts
    \item Monocytes
    \item NK cells
    \item Dendritic cells
    \item Neutrophils
    \item Basophils
}

Samples were additionally annotated to 29 fine cell types:
\itemize{
    \item Tcell.CD8.naive
    \item Tcell.CD8.CentralMemory
    \item Tcell.CD8.EffectorMemory
    \item Tcell.CD8.TerminalEffector
    \item Tcell.MAIT
    \item Tcell.VD2pos.gd
    \item Tcell.VD2neg.gd
    \item Tcell.FollicularHelper
    \item Tcell.Treg
    \item Tcell.Th1
    \item Tcell.Th1.Th17
    \item Tcell.Th17
    \item Tcell.Th2
    \item Tcell.CD4.naive
    \item Tcell.CD4.TerminalEffector
    \item Progenitor
    \item Bcell.naive
    \item Bcell.NonSwitchedMemory
    \item Bcell.Exhausted
    \item Bcell.SwitchedMemory
    \item Plasmablast
    \item Monocyte.Classical
    \item Monocyte.Int
    \item Monocyte.NonClassical
    \item NK
    \item Dendritic.Plasmacytoid
    \item Dendritic.Myeloid
    \item Neutrophil
    \item Basophil
}
}
\examples{
ref.se <- MonacoImmuneData()

}
\references{
Monaco G et al. (2019).
RNA-Seq Signatures Normalized by mRNA Abundance Allow Absolute Deconvolution of Human Immune Cell Types
\emph{Cell Rep.} 26, 1627-1640.
}
\author{
Jared Andrews
}
