% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleR.R
\docType{data}
\name{mouse.rnaseq}
\alias{mouse.rnaseq}
\title{Mouse-RNAseq reference dataset for mouse}
\format{A list with the following items:
\describe{
  \item{data}{Gene expression data matrix}
  \item{types}{vector of annotation per column in the matrix}
  \item{main_types}{vector of broad annotations per column in the matrix}
  \item{name}{name if the reference set}
  \item{sd.thres}{threshold of the standard deviation for genes to use in SingleR}
  \item{de.genes.main}{list of lists of differentially expressed genes between every two cell types in main_types}
  \item{de.genes}{list of lists of differentially expressed genes between every two cell types in types}
}}
\usage{
mouse.rnaseq
}
\description{
Mouse-RNAseq reference dataset for mouse
}
\keyword{datasets}
