% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleR.R
\docType{data}
\name{mouse_lengths}
\alias{mouse_lengths}
\title{Length of mouse genes for TPM calculation}
\format{A named vector of gene lengths}
\usage{
mouse_lengths
}
\description{
Length of mouse genes for TPM calculation
}
\keyword{datasets}
