% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{MouseRNAseqData}
\alias{MouseRNAseqData}
\title{Obtain mouse bulk expression data of sorted cell populations (RNA-seq)}
\usage{
MouseRNAseqData()
}
\value{
A \linkS4class{SummarizedExperiment} object with a \code{"logcounts"} assay
containing the log-normalized expression values, along with cell type labels in the 
\code{\link{colData}}.
}
\description{
Download and cache the normalized expression values of 358 bulk RNA-seq samples
of sorted cell populations that can be found at GEO.
}
\details{
This dataset was contributed by the Benayoun Lab that identified, 
downloaded and processed data sets on GEO that corresponded to sorted cell
types (Benayoun et al., 2019).

The dataset entails 358 mouse RNA-seq samples annotated to 18 main cell types: 
Adipocytes, Astrocytes, B cells, Cardiomyocytes, Dendritic cells, Endothelial 
cells, Epithelial cells, Erythrocytes, Fibroblasts, Granulocytes, Hepatocytes,
Macrophages, Microglia, Monocytes, Neurons, NK cells, Oligodendrocytes, T cells.
}
\examples{
ref.se <- MouseRNAseqData()

}
\references{
Benayoun B et al. (2019).
Remodeling of epigenome and transcriptome landscapes with aging in mice reveals widespread induction of inflammatory responses.
\emph{Genome Res.} 29, 697-709.

Code at \url{https://github.com/BenayounLaboratory/Mouse_Aging_Epigenomics_2018/tree/master/FigureS7_CIBERSORT/RNAseq_datasets_for_Deconvolution/2017-01-18}
}
\author{
Friederike Dündar
}
