% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizations.R
\name{plotCellVsReference}
\alias{plotCellVsReference}
\title{Plot a cell versus a reference}
\usage{
plotCellVsReference(test, test.id, ref, ref.id,
  assay.type.test = "logcounts", assay.type.ref = "logcounts")
}
\arguments{
\item{test}{A numeric matrix of single-cell expression values where rows are genes and columns are cells.
Alternatively, a \linkS4class{SummarizedExperiment} object containing such a matrix.}

\item{test.id}{Integer scalar or string specifying the index or name of the target cell to use.}

\item{ref}{A numeric matrix of reference expression values (usually log-transformed, see \code{\link{trainSingleR}}), where rows are genes and columns are cells.
Alternatively, a \linkS4class{SummarizedExperiment} object containing such a matrix.}

\item{ref.id}{Integer scalar or string specifying the index or name of the reference cell/sample to use.}

\item{assay.type.test}{Integer scalar or string specifying the assay of \code{test} containing the relevant expression data.
Used if \code{test} is a \linkS4class{SummarizedExperiment}.}

\item{assay.type.ref}{Integer scalar or string specifying the assay of \code{ref} containing the relevant expression data.
Used if \code{ref} is a \linkS4class{SummarizedExperiment}.}
}
\value{
A \link[ggplot2]{ggplot} object containing a scatter plot of the cell against a reference.
}
\description{
Plot a single cell's expression profile against that of a reference sample across all genes.
}
\details{
This function allows a user to manually check how an individual cell compares to reference cells of the same or different type.
It generates a scatter plot of one cell's expression profile against a chosen reference where each point is an individual gene.
It also displays a linear regression fit and the value of Spearman's rank correlation coefficient.
}
\examples{
# Running the SingleR() example.
example(SingleR, echo=FALSE)
test <- scater::logNormCounts(test)

# Checking if cell#1 resembles reference-set cells of its assigned label.
pred$labels[1]
same.type <- grep(pred$labels[1], sce$label)

# Compare expression of target cell to a reference cell of the same type
plotCellVsReference(test, test.id = 1, ref = sce, ref.id = same.type[1])

# Compare expression of target cell to reference cells of a different type
diff.type <- seq_along(pred$labels)[-same.type]
plotCellVsReference(test, test.id = 1, ref = sce, ref.id = diff.type[1])

}
\author{
Daniel Bunis, based on code by Dvir Aran.
}
