% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDeltaDistribution.R
\name{plotDeltaDistribution}
\alias{plotDeltaDistribution}
\title{Plot delta distributions}
\usage{
plotDeltaDistribution(
  results,
  show = c("delta.med", "delta.next"),
  labels.use = colnames(results$scores),
  deltas.use = NULL,
  calls.use = 0,
  size = 2,
  ncol = 5,
  dots.on.top = TRUE,
  this.color = "#000000",
  pruned.color = "#E69F00",
  grid.vars = list()
)
}
\arguments{
\item{results}{A \linkS4class{DataFrame} containing the output from \code{\link{SingleR}}, 
\code{\link{classifySingleR}}, \code{\link{combineCommonResults}}, or \code{\link{combineRecomputedResults}}.}

\item{show}{String specifying whether to show the difference from the median (\code{"delta.med"}) 
or the difference from the next-best score (\code{"delta.next"}).}

\item{labels.use}{Character vector specifying the labels to show in the plot facets.
Defaults to all labels in \code{results}.}

\item{deltas.use}{Integer scalar or vector specifying the individual annotation result from which to take deltas.
This is only relevant for combined results, see Details.}

\item{calls.use}{Integer scalar or vector specifying the individual annotation result from which to take labels.
This is only relevant for combined results, see Details.}

\item{size}{Numeric scalar to set the size of the dots.}

\item{ncol}{Integer scalar to set the number of labels to display per row.}

\item{dots.on.top}{Logical scalar specifying whether cell dots should be plotted on top of the violin plots.}

\item{this.color}{String specifying the color for cells that were assigned to the label.}

\item{pruned.color}{String specifying the color for cells that were assigned to the label but pruned.}

\item{grid.vars}{Deprecated, use \code{\link{plotDeltaDistributions}} instead.}
}
\value{
If \code{deltas.use} specifies a single set of deltas,
a \link[ggplot2]{ggplot} object is returned showing the deltas in violin plots.

If \code{deltas.use} specifies multiple deltas for a combined result,
multiple ggplot objects are generated in a grid on the current graphics device.

If \code{delta.use} specifies multiple deltas and \code{grid.vars} is set to \code{NULL},
a list is returned containing the ggplot objects for manual display.
}
\description{
Plot the distribution of deltas 
(i.e., the gap between the assignment score for the assigned label and those of the remaining labels)
across cells assigned to each reference label.
}
\details{
This function creates jitter and violin plots showing the deltas for all cells across one or more labels.
The idea is to provide a visual diagnostic for the confidence of assignment of each cell to its label.
The \code{show} argument determines what values to show on the y-axis:
\itemize{
\item \code{"delta.med"}, the difference between the score of the assigned label and the median of all scores for each cell.
\item \code{"delta.next"}, the difference between best and second-best scores of each cell at the last round of fine-tuning.
}

If any fine-tuning was performed, the highest scoring label for an individual cell may not be its final label.
This may manifest as negative values when \code{show="delta.med"}.
\code{show="delta.next"} is guaranteed to be positive but may be overly stringent for references involving very similar labels.

Pruned calls are identified as \code{NA}s in the \code{pruned.labels} field in \code{results}.
Points corresponding to cells with pruned calls are colored by \code{pruned.color};
this can be disabled by setting \code{pruned.color=NA}.

For combined results (see \code{?"\link{combine-predictions}"}),
this function can show both the combined and individual deltas or labels.
This is done using the \code{deltas.use} and \code{labels.use} arguments,
entries of which refer to columns of \code{results$orig.results}.
\code{labels.use} is also allowed to contain zero, in which case the combined labels are used.
For example:
\itemize{
\item If we set \code{deltas.use=2} and \code{calls.use=1},
we will plot the deltas from the second individual reference against the labels faceted from the first reference.
\item If we set \code{deltas.use=1:2} and \code{calls.use=1:2},
we will plot the deltas from first and second references (in separate plots) faceted by their corresponding labels.
\item By default, the function will create a separate plot for each individual reference faceted by the combined labels. 
This is equivalent to \code{deltas.use=1:N} and \code{calls.use=0} for \code{N} individual references.
}
}
\examples{
example(SingleR, echo=FALSE)

# Showing the delta to the median:
plotDeltaDistribution(pred)

# Showing the delta to the next-highest score:
plotDeltaDistribution(pred, show = "delta.next")

# Multi-reference compatibility:
example(combineRecomputedResults, echo = FALSE)
plotDeltaDistribution(results = combined)

# To have plots output in a grid rather than as separate pages, provide,
# a list of inputs for gridExtra::grid.arrange() to 'grids.vars'.
plotDeltaDistribution(combined, grid.vars = list(ncol = 1))

# An empty list will use grid.arrange defaluts
plotDeltaDistribution(combined, grid.vars = list())

}
\seealso{
\code{\link{SingleR}}, to generate scores.

\code{\link{pruneScores}}, to remove low-quality labels based on the scores, and to see more about the quailty cutoffs.

\code{\link[gridExtra]{grid.arrange}}, for tweaks to the how plots are arranged when multiple are output together.
}
\author{
Daniel Bunis and Aaron Lun
}
