% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotScoresByCellOrLabel.R
\name{plotScoresByCellOrLabel}
\alias{plotScoresByCellOrLabel}
\alias{plotScoresSingleCell}
\alias{plotScoresSingleLabel}
\alias{plotScoresMultiLabels}
\title{Plot SingleR scores on a per-label or per-cell basis.}
\usage{
plotScoresSingleCell(results, cell.id,
  labels.use = levels(as.factor(results$labels)), size = 2,
  colors = c("#F0E442", "#56B4E9", "gray70", "gray40"))

plotScoresSingleLabel(results, label, size = 0.5, dots.on.top = FALSE,
  colors = c("#F0E442", "#56B4E9", "gray70", "gray40"))

plotScoresMultiLabels(results, size = 0.2, dots.on.top = FALSE,
  labels.use = levels(as.factor(results$labels)), ncol = 5,
  colors = c("#F0E442", "#56B4E9", "gray70", "gray40"))
}
\arguments{
\item{results}{A \linkS4class{DataFrame} containing the output from \code{\link{SingleR}} or \code{\link{classifySingleR}}.}

\item{cell.id}{Integer specifying which cell to show for \code{plotScoresSingleCell}}

\item{labels.use}{String vector indicating what labels to show in \code{plotScoresSingleCell} and \code{plotScoresMultiLabel}
If \code{labels.use} is left \code{NULL}, all labels available in \code{results} are presented.}

\item{size}{Scalar, the size of the dots}

\item{colors}{String vector that sets the colors.  Order of colors should be: `this label`, `this label - pruned`, `other label`, `other label - pruned`.}

\item{label}{String indicating which individual label to plot in \code{plotScoresSingleLabel}
This input will be unnecessary once \code{\link{pruneScores}}'s output is added to the results DataFrame}

\item{dots.on.top}{Logical which sets whether cell dots are plotted on top of, versus behind, the violin plots in \code{plotScoresSingleLabel} and \code{plotScoresMultiLabel}}

\item{ncol}{Integer number of labels to display per row}
}
\value{
Each function returns a \link{ggplot} object showing SingleR scores in a dot and/or violin plot representation.
}
\description{
Plot SingleR scores on a per-label or per-cell basis.
}
\details{
The \code{plotScoresSingleCell} function creates a dot plot showing the scores of a single cell across many labels.
The score for the final label of an individual cell is marked in yellow, with a dotted line indicating the median score across all labels.
This function may be useful for visualizing and tuning the \code{min.diff.med} per-cell cutoff of the \code{\link{pruneScores}} function.

The \code{plotScoresSingleLabel} and \code{plotScoresMultiLabel} functions create jitter and violin plots showing
the scores of all cells across a single label or multiple labels, respectively.
For a given label X, cells are split into several categories:
\itemize{
\item Was assigned to label X, and the label was not pruned away.
\item Was assigned to label X, and the label was pruned away.
\item Was not assigned to label X, and the label was not pruned away.
\item Was not assigned to label X, and the label was pruned away.
}
Each category is grouped and colored separately.
These functions can be used to assess the distribution of scores of all cells for individual labels,
and may be useful for visualizing and tuning the \code{nmads} per-label cutoff of the \code{\link{pruneScores}} function.

Note that these functions show initial scores only, i.e., prior to fine tuning.
However, the labels may be defined after fine-tuning in \code{\link{SingleR}} or \code{\link{classifySingleR}}.
Thus, the best score for an individual cell may not be its final label.
}
\examples{
example(SingleR, echo=FALSE)

plotScoresSingleCell(results = pred, cell.id = 1)
plotScoresSingleLabel(results = pred, label = "B",
    dots.on.top = TRUE)
plotScoresMultiLabels(results = pred,
    dots.on.top = TRUE, size = 0.5)

}
\seealso{
\code{\link{SingleR}}, to generate scores.

\code{\link{pruneScores}}, to remove low-quality labels based on the scores.
}
\author{
Daniel Bunis
}
