% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{quantileMatrix}
\alias{quantileMatrix}
\title{Calculate a quantile value by groups}
\usage{
quantileMatrix(mat, groups, q)
}
\arguments{
\item{mat}{a matrix of numeric values NxM}

\item{groups}{a vector of size M with annotation of columns in mat}

\item{q}{the quantile to calculate per group}
}
\value{
a matrix with one column per group
}
\description{
Calculate a quantile value by groups
}
