% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleR.R
\name{ReadSingleCellData}
\alias{ReadSingleCellData}
\title{Internal function to read single-cell data}
\usage{
ReadSingleCellData(counts, annot)
}
\arguments{
\item{counts}{a tab delimited text file containing the counts matrix, a 10X directory name or a matrix with the counts.}

\item{annot}{a tab delimited text file or a data.frame. Rownames correspond to column names in the counts data}
}
\value{
list with sc.data - the counts matrix, orig.ident - a named vector of the originial identities.
}
\description{
Internal function to read single-cell data
}
