% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleR.Create.R
\name{remove.Unnecessary.Data.single}
\alias{remove.Unnecessary.Data.single}
\title{Remove data from a SingleR object to make it smaller}
\usage{
remove.Unnecessary.Data.single(singler.data)
}
\arguments{
\item{singler.data}{a SingleR object}
}
\value{
a smaller SingleR object
}
\description{
Remove data from a SingleR object to make it smaller
}
