% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleR.Object.R
\docType{class}
\name{SingleR-class}
\alias{SingleR-class}
\title{An S4 class to represent SingleR objects
This object can be upload to the SingleR browser (http)}
\description{
An S4 class to represent SingleR objects
This object can be upload to the SingleR browser (http)
}
\section{Slots}{

\describe{
\item{\code{project.name}}{a string. Must be the same as the file name.}

\item{\code{xy}}{matrix of the xy coordinates. Row names must be single-cell names.}

\item{\code{labels}}{data.frame of the SingleR annotations. Rows are single-cell, a column for each 
annotation by reference. Column names are the reference name (e.g. Immgen, Immgen.main, Blueprint+Encode, etc.)}

\item{\code{labels.NFT}}{Same as labels. Annotation before fine-tuning.}

\item{\code{labels.clusters}}{data.frame. Same as labels, but a row for each cluster, not single-cell.}

\item{\code{labels.clusters.NFT}}{Same as labels.clusters. Annotation before fine-tuning.}

\item{\code{clusters}}{data.frame. Rows are single-cell, columns are clusters. Multiple clsutering are possible.}

\item{\code{ident}}{data.frame. Rows are single-cell, columns are caterogical meta.data (e.g. orig.ident).}

\item{\code{other}}{data.frame. Rows are single-cell, columns are continous variables (e.g. cell cycle score).}

\item{\code{scores}}{list. Each item in the list corresponds to the column labels in labels. Items are the SingleR score matrices.}

\item{\code{meta.data}}{character array. Each item is a named charachter (e.g. Citation, Organism, Technology)}

\item{\code{expr}}{a matrix. The single-cell expression matrix.}
}}

