% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleR.R
\name{SingleR.Cluster}
\alias{SingleR.Cluster}
\title{SingleR.Cluster - using the scores data to cluster single cells.}
\usage{
SingleR.Cluster(SingleR, num.clusts = 10)
}
\arguments{
\item{SingleR}{the output from the SingleR function}

\item{num.clusts}{number of clusters to output}
}
\value{
cluster id for each single cell
}
\description{
SingleR.Cluster - using the scores data to cluster single cells.
}
