% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleR.Create.R
\name{SingleR.Combine}
\alias{SingleR.Combine}
\title{Combining SingleR objects together to one object.}
\usage{
SingleR.Combine(singler.list, order = NULL, clusters = NULL, expr = NULL,
  xy = NULL)
}
\arguments{
\item{singler.list}{a list of SingleR objects}

\item{order}{a vector of the cells name to order the SingleR object by. Important if there is a pre-built single-cell object, and the order of cells in SingleR needs to be in the same order.}

\item{clusters}{a vector of clusters  per cell. Optional.}

\item{expr}{the counts matrix. Optional.}

\item{xy}{the t-SNE coordinates of all cells.}
}
\value{
a combined SingleR object
}
\description{
This function can be useful when the memory is insufficient to analyze the whole counts matrix. We can analyze subsets of the full data and combine them together afterwards.
}
