% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleR.R
\name{SingleR.ConfidenceTest}
\alias{SingleR.ConfidenceTest}
\title{Test confidence of SingleR annotation.}
\usage{
SingleR.ConfidenceTest(scores)
}
\arguments{
\item{scores}{a SingleR scores matrix}
}
\value{
vector of p-values
}
\description{
Calculates outliers chisquare-test for the cell typres scores for each cell.
}
