% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleR.R
\name{SingleR.CreateFromCounts}
\alias{SingleR.CreateFromCounts}
\title{Wrapper function to create a SingleR+Seurat object}
\usage{
SingleR.CreateFromCounts(counts, annot = NULL, project.name,
  min.genes = 500, min.cells = 2, regress.out = "nUMI", npca = 10,
  technology = "10X", species = "Human", citation = "",
  ref.list = list(), reduce.file.size = T, normalize.gene.length = F,
  do.signatures = T, temp.dir = NULL)
}
\arguments{
\item{temp.dir}{}
}
\description{
Wrapper function to create a SingleR+Seurat object
}
