% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleR.R
\name{SingleR.CreateKangAnnotations}
\alias{SingleR.CreateKangAnnotations}
\title{Run annotation based on Kang et al. Nature Biotechnology 2017.}
\usage{
SingleR.CreateKangAnnotations(sc.data)
}
\arguments{
\item{sc.data}{a matrix of single cell expression data.}
}
\value{
list of the correlation matrix and the annotations.
}
\description{
Run annotation based on Kang et al. Nature Biotechnology 2017.
}
