% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleR.R
\name{SingleR.CreateSeurat}
\alias{SingleR.CreateSeurat}
\title{Wrapper function to create a Seurat object}
\usage{
SingleR.CreateSeurat(project, sc.data, min.genes = 500, min.cells = 2,
  regress.out = "nUMI", npca = 10, resolution = 0.8, species = "Human",
  temp.dir = NULL)
}
\arguments{
\item{temp.dir}{}
}
\description{
Wrapper function to create a Seurat object
}
