% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SinlgeR.Plotting.R
\name{SingleR.DrawBoxPlot}
\alias{SingleR.DrawBoxPlot}
\title{Plot boxplots for each label for a given single cell.}
\usage{
SingleR.DrawBoxPlot(sc_data, cell_id, ref, labels.use = NULL,
  quantile.order = 0.8, main_types = F, top.n = 50, tit = NULL,
  colors = singler.colors)
}
\arguments{
\item{sc_data}{the single-cell RNA-seq data set as a matrix with genes as rownames.}

\item{cell_id}{a number of the single cell to use}

\item{ref}{the reference dataset with genes as rownames. Gene names must be in the same format as the single cell data (if sc_data uses genes symbols, ref_data must have the same)}

\item{labels.use}{a list of labels to use. If NULL uses all labels.}

\item{quantile.order}{same a quantile.use - by which percentile to order the boxplots.}

\item{main_types}{aggregate labels by main types or by all types}

\item{top.n}{number of boxplots to present (starting from top)}

\item{tit}{title for the boxplot}

\item{colors}{colors to use. Default is singler.colors}
}
\value{
a list with a ggplot and the scores for the single cell
}
\description{
Plot boxplots for each label for a given single cell.
}
