% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleR.Plotting.R
\name{SingleR.DrawScatter}
\alias{SingleR.DrawScatter}
\title{Plot a scatter plot of a single cell vs. a reference sample}
\usage{
SingleR.DrawScatter(sc_data, cell_id, ref, sample_id)
}
\arguments{
\item{sc_data}{the single-cell RNA-seq data set as a matrix with genes as rownames.}

\item{cell_id}{a number of the single cell to use}

\item{ref}{the reference dataset with genes as rownames. Gene names must be in the same format as the single cell data (if sc_data uses genes symbols, ref_data must have the same)}

\item{sample_id}{a number of the sample to use}
}
\value{
a ggplot
}
\description{
Plot a scatter plot of a single cell vs. a reference sample
}
