% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleR.Plotting.R
\name{SingleR.PlotFeature}
\alias{SingleR.PlotFeature}
\title{Plot a feature on the t-SNE plot}
\usage{
SingleR.PlotFeature(SingleR, seurat, plot.feature = "MaxScore",
  dot.size = 1, title = NULL)
}
\arguments{
\item{SingleR}{the output from the SingleR function}

\item{seurat}{the seurat object}

\item{plot.feature}{the feature to plot - 'MaxScore' for coloring according to the top score per cell, 'nGene' for number of non-zero genes, 'nUMI' for number of UMIs, or a gene name.}

\item{dot.size}{size of the dot in the plot.}
}
\value{
ggplot2 object
}
\description{
Plot a feature on the t-SNE plot
}
