% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleR.R
\name{SingleR.PlotTsne}
\alias{SingleR.PlotTsne}
\title{Plot a colored tSNE plot according to a set of labels}
\usage{
SingleR.PlotTsne(SingleR, xy, labels = SingleR$labels, clusters = NULL,
  do.letters = TRUE, dot.size = 1, do.labels = FALSE, do.legend = TRUE,
  label.size = 3, title = "", colors = singler.colors, font.size = NULL,
  alpha = 0.5)
}
\arguments{
\item{SingleR}{the output from the SingleR function}

\item{xy}{a matrix with the coordinates of the single cells}

\item{labels}{labels for the single cells, defualt is the SingleR labels}

\item{clusters}{if the SingleR method is 'clusters' then this vector is the cluster id for each single cell}

\item{do.letters}{if TRUE shows letters the first letter of the annotation}

\item{dot.size}{the size of the dot. Default is 1}

\item{do.labels}{only relevant in a 'cluster' method. If TRUE The cluster annotation is shown on top of the cluster.}

\item{do.legend}{if TRUE legend is presened.}

\item{label.size}{if do.labels is TRUE then this determines the font size of the labels}

\item{title}{title for the plot}

\item{colors}{colors to use}

\item{font.size}{size of fonts in the plot}

\item{alpha}{an alpha for the transperancy of the dots}
}
\value{
a list with ggplot and a data frame with the coordinates and annotations
}
\description{
Plot a colored tSNE plot according to a set of labels
}
