% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleR.R
\name{superRef}
\alias{superRef}
\title{Creating a subset of the super reference dataset}
\usage{
superRef(ref.name, ref, ids)
}
\arguments{
\item{ref.name}{name of the reference dataset}

\item{ref}{the super reference dataset}

\item{ids}{numeric vector of indexes to use}
}
\value{
a specfic super reference dataset
}
\description{
Creating a subset of the super reference dataset
}
