% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{TPM}
\alias{TPM}
\title{Normalize counts to gene length}
\usage{
TPM(counts, lengths = NULL)
}
\arguments{
\item{counts}{counts data}

\item{lengths}{genes' lengths. If null use a precalculated gene lengths}
}
\value{
TPM values
}
\description{
Normalize counts to gene length
}
