% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleR.Object.R
\name{convertSingleR2Browser}
\alias{convertSingleR2Browser}
\title{Convert the SingleR output object to a SingleR S4 object that can be uploaded to the SingleR browser.}
\usage{
convertSingleR2Browser(singler)
}
\arguments{
\item{singler}{the SingleR output object}
}
\value{
S4 SingleR object that can be uploaded to the SingleR browser
}
\description{
Convert the SingleR output object to a SingleR S4 object that can be uploaded to the SingleR browser.
}
