% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{DatabaseImmuneCellExpressionData}
\alias{DatabaseImmuneCellExpressionData}
\title{Obtain human bulk RNA-seq data from the Database of Immune Cell Expression (DICE)}
\usage{
DatabaseImmuneCellExpressionData()
}
\value{
A \linkS4class{SummarizedExperiment} object with a \code{"logcounts"} assay
containing the log-normalized expression values, along with cell type labels in the 
\code{\link{colData}}.
}
\description{
Download and cache the normalized expression values of 1561 bulk RNA-seq samples
of sorted cell populations from the DICE database.
}
\details{
This function provides normalized expression values of 1561 bulk RNA-seq samples
generated by DICE from pure populations of human immune cells (<http://www.dice-database.org/>).

TPM normalized values for each cell type were downloaded from \url{https://dice-database.org/downloads}. 
Genes with no reads across samples were removed, and values were log2 normalized after a pseudocount of 1 was added.

The dataset entails 1561 human RNA-seq samples annotated to 5 main cell types: 
"B cells", "Monocytes", "NK cells", "T cells, CD8+", and "T cells, CD4+".

Samples were additionally annotated to 15 fine cell types:
"B cells, naive", "Monocytes, CD14+", "Monocytes, CD16+", "NK cells", "T cells, memory TREG",
"T cells, CD4+, naive", "T cells, CD4+, naive, stimulated", "T cells, CD4+, naive Treg",
"T cells, CD4+, Th1", "T cells, CD4+, Th1_17", "T cells, CD4+, Th2", "T cells, CD8+, naïve",
"T cells, CD8+, naïve, stimulated", "T cells, CD4+, TFH", "T cells, CD4+, Th17".
}
\examples{
ref.se <- DatabaseImmuneCellExpressionData()

}
\references{
Schmiedel B et al. (2018).
Impact of Genetic Polymorphisms on Human Immune Cell Gene Expression.
\emph{Cell} 175, 1701-1715.
}
\author{
Jared Andrews
}
