% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{DatabaseImmuneCellExpressionData}
\alias{DatabaseImmuneCellExpressionData}
\title{Obtain human bulk RNA-seq data from the Database of Immune Cell Expression (DICE)}
\usage{
DatabaseImmuneCellExpressionData()
}
\value{
A \linkS4class{SummarizedExperiment} object with a \code{"logcounts"} assay
containing the log-normalized expression values, along with cell type labels in the 
\code{\link{colData}}.
}
\description{
Download and cache the normalized expression values of 1561 bulk RNA-seq samples
of sorted cell populations from the DICE database.
}
\details{
This function provides normalized expression values of 1561 bulk RNA-seq samples
generated by DICE from pure populations of human immune cells (<http://www.dice-database.org/>).

TPM normalized values for each cell type were downloaded from \url{https://dice-database.org/downloads}. 
Genes with no reads across samples were removed, and values were log2 normalized after a pseudocount of 1 was added.

The dataset entails 1561 human RNA-seq samples annotated to 5 main cell types: 
\itemize{
    \item B cells
    \item Monocytes
    \item NK cells
    \item T cells, CD8+
    \item T cells, CD4+
}

Samples were additionally annotated to 15 fine cell types:
\itemize{
    \item B cells, naive
    \item Monocytes, CD14+
    \item Monocytes, CD16+
    \item NK cells
    \item T cells, memory TREG
    \item T cells, CD4+, naive
    \item T cells, CD4+, naive, stimulated
    \item T cells, CD4+, naive Treg
    \item T cells, CD4+, Th1
    \item T cells, CD4+, Th1_17
    \item T cells, CD4+, Th2
    \item T cells, CD8+, naïve
    \item T cells, CD8+, naïve, stimulated
    \item T cells, CD4+, TFH
    \item T cells, CD4+, Th17
}
}
\examples{
ref.se <- DatabaseImmuneCellExpressionData()

}
\references{
Schmiedel B et al. (2018).
Impact of Genetic Polymorphisms on Human Immune Cell Gene Expression.
\emph{Cell} 175, 1701-1715.
}
\author{
Jared Andrews
}
