% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{ImmGenData}
\alias{ImmGenData}
\title{Obtain mouse bulk expression data from the Immunologic Genome Project}
\usage{
ImmGenData()
}
\value{
A \linkS4class{SummarizedExperiment} object with a \code{"normcounts"} assay
containing the normalized expression values and cell type labels in the \code{\link{colData}}.
}
\description{
Download and cache the normalized expression values of 830 microarray samples of
pure mouse immune cells, generated by the Immunologic Genome Project (ImmGen).
}
\details{
This function provides normalized expression values of 830 microarray samples
generated by ImmGen from pure populations of murine immune cells (<http://www.immgen.org/>).

The samples were processed and normalized as described in Aran, Looney &
Liu et al. (2019); i.e. CEL files from the Gene Expression Omnibus (GEO; GSE15907 and GSE37448), 
were downloaded, processed, and normalized using the robust multi-array average
(RMA) procedure on probe-level data (with Matlab functions).

The data will be downloaded from ExperimentHub,
returning a \linkS4class{SummarizedExperiment} object for further use.
}
\examples{
ref.se <- ImmGenData()

}
\references{
The Immunological Genome Project (2008).
Networks of gene expression in immune cells.
\emph{Nature Immunology}. doi: 10.1038/ni1008-1091

Processing described in Aran, Looney & Liu et al. (2019). 
\emph{Nature Immunology} 20, 163–172. doi: 10.1038/s41590-018-0276-y
}
\author{
Friederike Duendar
}
