% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{NovershternHematopoieticData}
\alias{NovershternHematopoieticData}
\title{Obtain human bulk microarray expression data of sorted cell populations from GSE24759 (DMAP)}
\usage{
NovershternHematopoieticData()
}
\value{
A \linkS4class{SummarizedExperiment} object with a \code{"logcounts"} assay
containing the log-normalized expression values, along with cell type labels in the 
\code{\link{colData}}.
}
\description{
Download and cache the normalized expression values of 211 bulk human microarray samples
of sorted hematopoietic cell populations that can be found in 
\href{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE24759}{GSE24759}.
}
\details{
The dataset entails 211 human microarray samples annotated to 17 main cell types: 
Basophils, B cells, CMPs, Dendritic cells, Eosinophils, Erythroid cells, GMPS,
Granulocytes, HSCs, Megakaryocytes, MEPs, Monocytes, NK cells, NK T cells,,
CD8+ T cells, CD4+ T cells.

Samples were additionally annotated to 38 fine cell types:
Basophils, Naive B cells, Mature B-cells class able to switch, Mature B-cells,
Mature B-cells class switched, Common myeloid progenitors, Plasmacytoid Dendritic Cells,
Myeloid Dendritic Cells, Eosinophils, Erythroid_CD34+ CD71+ GlyA-, Erythroid_CD34- CD71+ GlyA-,
Erythroid_CD34- CD71+ GlyA+, Erythroid_CD34- CD71lo GlyA+, Erythroid_CD34- CD71- GlyA+, 
Granulocyte/monocyte progenitors, Colony Forming Unit-Granulocytes, 
Granulocyte (Neutrophilic Metamyelocytes), Granulocyte (Neutrophils), 
Hematopoietic stem cells_CD133+ CD34dim, Hematopoietic stem cell_CD38- CD34+,
Colony Forming Unit-Megakaryocytic, Megakaryocytes, Megakaryocyte/erythroid progenitors,
Colony Forming Unit-Monocytes, Monocytes, Mature NK cells_CD56- CD16+ CD3-,
Mature NK cells_CD56+ CD16+ CD3-, Mature NK cells_CD56- CD16- CD3-, NK T cells,
Early B-cells, Pro B-cells, CD8+ Effector Memory RA, Naive CD8+ T-cells, 
CD8+ Effector Memory, CD8+ Central Memory, Naive CD4+ T-cells, CD4+ Effector Memory,
CD4+ Central Memory.
}
\examples{
ref.se <- NovershternHematopoieticData()

}
\references{
Novershtern N et al. (2011).
Densely interconnected transcriptional circuits control cell states in human hematopoiesis.
\emph{Cell} 144, 296-309.
}
\author{
Jared Andrews
}
