% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizations.R
\name{plotCellVsReference}
\alias{plotCellVsReference}
\title{Plot expression of a cell versus a reference cell/sample}
\usage{
plotCellVsReference(sc.data, sc.id, train.data, train.id,
  assay.type.sc = "logcounts", assay.type.train = "logcounts")
}
\arguments{
\item{sc.data}{Numeric matrix of single-cell expression values (usually log-transformed
or otherwise variance-stabilized), where rows are genes and columns are cells.
Alternatively, a \linkS4class{SingleCellExperiment} object containing such a matrix.}

\item{sc.id}{Integer scalar specifying the index of the target cell to use.
Alternatively, a string containing the name of the cell.}

\item{train.data}{Numeric matrix of reference dataset expression values (usually log-transformed
or otherwise variance-stabilized), where rows are genes and columns are cells.
Alternatively, a \linkS4class{SingleCellExperiment} object containing such a matrix.}

\item{train.id}{Integer scalar specifying the reference cell/sample to use.}

\item{assay.type.sc}{Integer scalar or string specifying the assay of \code{sc.data} containing the relevant expression data.
Used if \code{sc.data} is a \linkS4class{SingleCellExperiment}.}

\item{assay.type.train}{Integer scalar or string specifying the assay of \code{train.data} containing the relevant expression data.
Used if provided \code{train.data} is a \linkS4class{SingleCellExperiment}.}
}
\value{
A \link{ggplot} object containing a scatter plot of the cell against a reference.
}
\description{
Plot expression of a cell versus a reference cell/sample
}
\details{
This function allows a user to manually check how an individual cell compares to reference cells of the same or different type.
It generates a scatter plot of one cell's expression profile against a chosen reference where each point is an individual gene.
It also displays a linear regression fit and the value of Spearman's rank correlation coefficient.
}
\examples{
# Running the SingleR() example.
example(SingleR, echo=FALSE)
test <- scater::logNormCounts(test)

# Checking if cell#1 resembles reference-set cells of its assigned label.
pred$labels[1]
same.type <- grep(pred$labels[1], sce$label)

# Compare expression of target cell to a reference cell of the same type
plotCellVsReference(test, sc.id = 1, train.data = sce, train.id = same.type[1])

# Compare expression of target cell to reference cells of a different type
diff.type <- seq_along(pred$labels)[-ref.sameType.as1]
plotCellVsReference(test, sc.id = 1, train.data = sce, train.id = diff.type[1])

}
\author{
Daniel Bunis, based on code by Dvir Aran.
}
