% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotScoreDistribution.R
\name{plotScoreDistribution}
\alias{plotScoreDistribution}
\title{Plot score distributions of labels.}
\usage{
plotScoreDistribution(results, show = c("delta.med", "delta.next",
  "scores"), labels = colnames(results$scores), size = 0.5, ncol = 5,
  dots.on.top = TRUE, this.color = "#F0E442",
  pruned.color = "#E69F00", other.color = "gray60",
  show.nmads = NULL, show.min.diff = NULL)
}
\arguments{
\item{results}{A \linkS4class{DataFrame} containing the output from \code{\link{SingleR}} or \code{\link{classifySingleR}}.}

\item{show}{String specifying whether to show the scores, the difference from the median or the difference from the next-best score.}

\item{labels}{String vector indicating one or more labels to show.
If \code{NULL}, all labels available in \code{results} are presented.}

\item{size}{Numeric scalar to set the size of the dots.}

\item{ncol}{Integer scalar to set the number of labels to display per row.}

\item{dots.on.top}{Logical specifying whether cell dots should be plotted on top of the violin plots.}

\item{this.color}{String specifying the color for cells that were assigned to the label.}

\item{pruned.color}{String specifying the color for cells that were assigned to the label but pruned.}

\item{other.color}{String specifying the color for other cells not assigned to the label.}

\item{show.nmads}{Numeric scalar that shows the threshold that would be used for pruning with \code{\link{pruneScores}}.
Only used when \code{show="delta.med"}.}

\item{show.min.diff}{Numeric scalar that shows the threshold that would be used for pruning with \code{\link{pruneScores}}.
Only used when \code{show="delta.med"} or \code{"delta.next"}.}
}
\value{
A \link[ggplot2]{ggplot} object showing assignment scores in violin plots.
}
\description{
Plot score distributions of labels.
}
\details{
This function creates jitter and violin plots showing assignment scores or related values for all cells across one or more labels.
It is intended for visualizing and adjusting the \code{nmads}, \code{min.diff.med}, and \code{min.diff.next} cutoffs of the \code{\link{pruneScores}} function.
 
The \code{show} argument determines what values to show on the y-axis.
Options are:
\itemize{
\item \code{"delta.med"}, the difference between the score of the assigned label and the median of all scores for each cell.
\item \code{"delta.next"}, the difference between best and second-best tuning scores of each cell.
\item \code{"scores"}, the raw assignment scores prior to fine-tuning.
}

For a given label X, cells distributions in several categories are shown:
\itemize{
\item Was assigned to label X, and the label was not pruned away.
\item Was assigned to label X, and the label was pruned away.
\item Was assigned as any label, including label X.
}
Each category is grouped and colored separately based on \code{this.color} and related parameters.

Values are stratified according to the assigned labels in \code{results$labels}.
If any fine-tuning was performed, the highest scoring label for an individual cell may not be its final label.
This may manifest as negative values when \code{show="delta.med"}.

Also note that \code{\link{pruneScores}} trims based on the \code{min.diff.med} and \code{min.diff.next} cutoffs first,
before calculating the first-labels' delta medians.
Thus, the actual \code{nmads} cut-off used in \code{\link{pruneScores}} may vary from the one portrayed in the plot.
}
\examples{
example(SingleR, echo=FALSE)

# To show the distribution of scores grouped by label:
plotScoreDistribution(results = pred)
# We can display a particular label using the label
plotScoreDistribution(results = pred, labels = "B")

# To show the distribution of deltas between cells' maximum and median scores,
#   grouped by label, change `show` to "delta.med":
#   This is useful for checking/adjusting nmads and min.diff.med
plotScoreDistribution(results = pred, show = "delta.med")
# The nmads cutoff can be displayed using show.nmads.
plotScoreDistribution(results = pred, show = "delta.med",
    show.nmads = 3)
# A min.diff.med cutoff can be shown using show.min.diff
plotScoreDistribution(results = pred, show = "delta.med",
    show.min.diff = 0.03)

# To show the distribution of deltas between cells' top 2 fine-tuning scores,
#   grouped by label, change `show` to "delta.next":
#   This is useful for checking/adjusting min.diff.next
plotScoreDistribution(results = pred, show = "delta.next")
# A min.diff.med cutoff can be shown using show.min.diff
plotScoreDistribution(results = pred, show = "delta.next",
    show.min.diff = 0.03)

}
\seealso{
\code{\link{SingleR}}, to generate scores.

\code{\link{pruneScores}}, to remove low-quality labels based on the scores, and to see more about the quailty cutoffs.
}
\author{
Daniel Bunis and Aaron Lun
}
