% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotScoreDistribution.R
\name{plotScoreDistribution}
\alias{plotScoreDistribution}
\title{Plot score distributions}
\usage{
plotScoreDistribution(
  results,
  show = c("scores", "delta.med", "delta.next"),
  labels.use = colnames(results$scores),
  scores.use = NULL,
  calls.use = 0,
  pruned.use = NULL,
  size = 0.5,
  ncol = 5,
  dots.on.top = TRUE,
  this.color = "#F0E442",
  pruned.color = "#E69F00",
  other.color = "gray60",
  show.nmads = 3,
  show.min.diff = NULL,
  grid.vars = list()
)
}
\arguments{
\item{results}{A \linkS4class{DataFrame} containing the output from \code{\link{SingleR}}, 
\code{\link{classifySingleR}}, \code{\link{combineCommonResults}}, or \code{\link{combineRecomputedResults}}.}

\item{show}{Deprecated, use \code{\link{plotDeltaDistributions}} instead for \code{show!="scores"}.}

\item{labels.use}{Character vector specifying the labels to show in the plot facets.
Defaults to all labels in \code{results}.}

\item{scores.use}{Integer scalar or vector specifying the individual annotation result from which to take scores.
This is only relevant for combined results, see Details.}

\item{calls.use}{Integer scalar or vector specifying the individual annotation result from which to take labels.
This is only relevant for combined results, see Details.}

\item{pruned.use}{Deprecated and ignored.}

\item{size}{Numeric scalar to set the size of the dots.}

\item{ncol}{Integer scalar to set the number of labels to display per row.}

\item{dots.on.top}{Logical scalar specifying whether cell dots should be plotted on top of the violin plots.}

\item{this.color}{String specifying the color for cells that were assigned to the label.}

\item{pruned.color}{String specifying the color for cells that were assigned to the label but pruned.}

\item{other.color}{String specifying the color for other cells not assigned to the label.}

\item{show.nmads, show.min.diff, grid.vars}{Deprecated, use \code{\link{plotDeltaDistributions}} instead.}

\item{grid.vars}{Named list of extra variables to pass to \code{\link[gridExtra]{grid.arrange}},
used to arrange the multiple plots generated when \code{scores.use} is of length greater than 1.}
}
\value{
If \code{scores.use} specifies a single set of scores,
a \link[ggplot2]{ggplot} object is returned showing the scores in violin plots.

If \code{scores.use} specifies multiple scores for a combined result,
multiple ggplot objects are generated in a grid on the current graphics device.

If \code{scores.use} specifies multiple scores and \code{grid.vars=NULL},
a list is returned containing the ggplot objects for manual display.
}
\description{
Plot the distribution of assignment scores across all cells assigned to each reference label.
}
\details{
This function creates jitter and violin plots showing assignment scores for all cells across one or more labels.
Each facet represents a label in \code{labels.use} and contains three violin plots:
\itemize{
\item \dQuote{Assigned}, containing scores for all cells assigned to that label.
Colored according to \code{this.color}.
\dQuote{Pruned}, containing scores for all cells assigned to that label but pruned out.
Colored according to \code{pruned.color}, and can be omitted by setting \code{pruned.color=NA}.
\item \dQuote{Other}, containing the scores for all cells assigned to other labels.
Colored according to \code{other.color}.
}
The expectation is that the former is higher than the latter,
though the deltas generated by \code{\link{plotDeltaDistribution}} are often more informative in this regard.

For combined results (see \code{?"\link{combine-predictions}"}),
this function can show both the combined and individual scores or labels.
This is done using the \code{scores.use} and \code{calls.use} arguments,
entries of which refer to columns of \code{results$orig.results} if positive or to the combined results if zero.
For example:
\itemize{
\item If we set \code{scores.use=2} and \code{calls.use=1},
we will plot the scores from the second individual reference faceted by the labels predicted from the first reference.
\item If we set \code{scores.use=1:2} and \code{calls.use=1:2},
we will plot the scores from first and second references (in separate plots) faceted by their corresponding labels.
\item By default, the function will create a separate plot for the combined scores and each individual reference,
where the scores in each plot are faceted by the combined labels.
This is equivalent to \code{scores.use=0:N} and \code{calls.use=0} for \code{N} individual references.
}
}
\examples{
example(SingleR, echo=FALSE)

# To show the distribution of scores grouped by label:
plotScoreDistribution(results = pred)

# We can display a particular label using the label
plotScoreDistribution(results = pred,
    labels.use = "B")

# When SingleR is run with multiple references, default output will contain
# separate plots for each original reference, as well as for the the combined
# set when 'show' = "scores".
example(combineRecomputedResults, echo = FALSE)
plotScoreDistribution(results = combined)

# 'scores.use' sets which original results to plot distributions for, and can
# be multiple or NULL (default)
plotScoreDistribution(results = combined, scores.use = 0)
plotScoreDistribution(results = combined, scores.use = 1:2)

# To have plots output in a grid rather than as separate pages, provide,
# a list of inputs for gridExtra::grid.arrange() to 'grids.vars'.
plotScoreDistribution(combined, grid.vars = list(ncol = 1))

# An empty list will use grid.arrange defaluts
plotScoreDistribution(combined, grid.vars = list())

}
\seealso{
\code{\link{SingleR}}, to generate scores.

\code{\link{pruneScores}}, to remove low-quality labels based on the scores, and to see more about the quailty cutoffs.

\code{\link[gridExtra]{grid.arrange}}, for tweaks to the how plots are arranged when multiple are output together.
}
\author{
Daniel Bunis and Aaron Lun
}
