% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotScoreDistribution.R
\name{plotScoreDistribution}
\alias{plotScoreDistribution}
\title{Plot score distributions of labels.}
\usage{
plotScoreDistribution(
  results,
  show = c("scores", "delta.med", "delta.next"),
  labels.use = colnames(results$scores),
  scores.use = NULL,
  calls.use = 0,
  pruned.use = 0,
  size = 0.5,
  ncol = 5,
  dots.on.top = TRUE,
  this.color = "#F0E442",
  pruned.color = "#E69F00",
  other.color = "gray60",
  show.nmads = 3,
  show.min.diff = NULL,
  grid.vars = list()
)
}
\arguments{
\item{results}{A \linkS4class{DataFrame} containing the output from \code{\link{SingleR}} or \code{\link{classifySingleR}}.}

\item{show}{String specifying whether to show the scores, the difference from the median or the difference from the next-best score.}

\item{labels.use}{String vector indicating one or more labels to show.
If \code{NULL}, all labels available in \code{results} are presented.}

\item{scores.use}{Integer scalar specifying which scores to use.
However, this can also refer to any column index of \code{results$orig.results} to use scores from individual references in a combined prediction (see \code{?\link{combine-predictions}}.

Alternatively, \code{scores.use} can be an integer vector containing multiple such column indices (or zero).
In such cases, multiple plots will be created showing multiple sets of scores.

Default setting, \code{scores.use=NULL}, will create plots for all targets that make sense:
when \code{show="scores"} plots are created for top-level results & individual references;
when \code{show="delta.med"} or \code{"delta.next"}, plots are created for individual references;}

\item{calls.use, pruned.use}{Integer scalar specifying which labels or pruning calls to use, defaulting to those from the top-level \code{results}.
However, this can also refer to any column index of \code{results$orig.results} to use labels from individual references in a combined prediction (see \code{?\link{combine-predictions}}.

Alternatively, an integer vector of the same length as \code{scores.use},
specifying the labels to use in each plot generated by \code{scores.use}.}

\item{size}{Numeric scalar to set the size of the dots.}

\item{ncol}{Integer scalar to set the number of labels to display per row.}

\item{dots.on.top}{Logical specifying whether cell dots should be plotted on top of the violin plots.}

\item{this.color}{String specifying the color for cells that were assigned to the label.}

\item{pruned.color}{String specifying the color for cells that were assigned to the label but pruned.}

\item{other.color}{String specifying the color for other cells not assigned to the label.}

\item{show.nmads}{Numeric scalar that shows the threshold that would be used for pruning with \code{\link{pruneScores}}.
Only used when \code{show="delta.med"}.}

\item{show.min.diff}{Numeric scalar that shows the threshold that would be used for pruning with \code{\link{pruneScores}}.
Only used when \code{show="delta.med"} or \code{"delta.next"}.}

\item{grid.vars}{named list of extra variables to pass to \code{\link[gridExtra]{grid.arrange}},
used when \code{scores.use} is of length greater than 1.
If \code{NULL}, the function will not arrange plots on a grid but simply create them one after another on the graphics device.}
}
\value{
One or more \link[ggplot2]{ggplot} objects showing assignment scores in violin plots is generated on the current graphics device.
A list of such objects is returned if \code{scores.use} is of length greater than 1.
}
\description{
Plot score distributions of labels.
}
\details{
This function creates jitter and violin plots showing assignment scores or related values for all cells across one or more labels.
It is intended for visualizing and adjusting the \code{nmads}, \code{min.diff.med}, and \code{min.diff.next} cutoffs of the \code{\link{pruneScores}} function.

The \code{show} argument determines what values to show on the y-axis.
Options are:
\itemize{
\item \code{"delta.med"}, the difference between the score of the assigned label and the median of all scores for each cell.
\item \code{"delta.next"}, the difference between best and second-best tuning scores of each cell.
\item \code{"scores"}, the raw assignment scores prior to fine-tuning.
}

For a given label X, cells distributions in several categories are shown:
\itemize{
\item Was assigned to label X, and the label was not pruned away.
\item Was assigned to label X, and the label was pruned away.
\item Was assigned as any label, including label X.
}
Each category is grouped and colored separately based on \code{this.color} and related parameters.

Values are stratified according to the assigned labels in \code{results$labels}.
If any fine-tuning was performed, the highest scoring label for an individual cell may not be its final label.
This may manifest as negative values when \code{show="delta.med"}.

Also note that \code{\link{pruneScores}} trims based on the \code{min.diff.med} and \code{min.diff.next} cutoffs first,
before calculating the first-labels' delta medians.
Thus, the actual \code{nmads} cut-off used in \code{\link{pruneScores}} may vary from the one portrayed in the plot.
}
\examples{
example(SingleR, echo=FALSE)

# To show the distribution of scores grouped by label:
plotScoreDistribution(results = pred)
# We can display a particular label using the label
plotScoreDistribution(results = pred,
    labels.use = "B")

# To show the distribution of deltas between cells' maximum and median scores,
#   grouped by label, change 'show' to "delta.med":
#   This is useful for checking/adjusting nmads and min.diff.med cutoffs
plotScoreDistribution(results = pred,
    show = "delta.med")

# To show the distribution of deltas between cells' top 2 fine-tuning scores,
#   grouped by label, change `show` to "delta.next":
#   This is useful for checking/adjusting min.diff.next cutoffs
plotScoreDistribution(results = pred, show = "delta.next")


### Visualizing and adjusting pruning cutoffs ###

# The default nmads cutoff of 3 is displayed when 'show = "delta.med"', but
# this can be adjusted or turned off with 'show.nmads'
plotScoreDistribution(results = pred,
    show = "delta.med", show.nmads = 2)
plotScoreDistribution(results = pred,
    show = "delta.med", show.nmads = NULL)

# A min.diff cutoff can be shown using 'show.min.diff' when
# 'show = "delta.med"' or 'show = "delta.next"'
plotScoreDistribution(results = pred,
    show = "delta.med", show.min.diff = 0.03)
plotScoreDistribution(results = pred,
    show = "delta.next", show.min.diff = 0.03)


### Multi-Reference Compatibility ###

# When SingleR is run with multiple references, default output will contain
# separate plots for each original reference, as well as for the the combined
# set when 'show' = "scores".
example(combineRecomputedResults, echo = FALSE)
plotScoreDistribution(results = combined)

# 'scores.use' sets which original results to plot distributions for, and can
# be multiple or NULL (default)
plotScoreDistribution(results = combined, show = "scores",
    scores.use = 0)
plotScoreDistribution(results = combined, show = "scores",
    scores.use = 1:2)

# To color and group cells by non-final label and pruned calls,
# use 'calls.use' and 'pruned.use'
plotScoreDistribution(results = combined, show = "scores",
    calls.use = 1, pruned.use = 1)

# To have plots output in a grid rather than as separate pages, provide,
# a list of inputs for gridExtra::grid.arrange() to 'grids.vars'.
plotScoreDistribution(combined,
    grid.vars = list(ncol = 1))

# An empty list will use grid.arrange defaluts
plotScoreDistribution(combined,
    grid.vars = list())

}
\seealso{
\code{\link{SingleR}}, to generate scores.

\code{\link{pruneScores}}, to remove low-quality labels based on the scores, and to see more about the quailty cutoffs.
}
\author{
Daniel Bunis and Aaron Lun
}
