% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotScoreDistribution.R
\name{plotScoreDistribution}
\alias{plotScoreDistribution}
\title{Plot score distributions of labels.}
\usage{
plotScoreDistribution(results, show = "scores",
  labels = colnames(results$scores), size = 0.5, ncol = 5,
  dots.on.top = TRUE, colors = c("#F0E442", "#E69F00", "gray60"))
}
\arguments{
\item{results}{A \linkS4class{DataFrame} containing the output from \code{\link{SingleR}} or \code{\link{classifySingleR}}.}

\item{show}{A string selecting the scores representation to show.}

\item{labels}{String vector indicating one or more labels to show.
If \code{labels} is left \code{NULL}, all labels available in \code{results} are presented.}

\item{size}{Scalar which sets the size of the dots}

\item{ncol}{Integer number of labels to display per row}

\item{dots.on.top}{Logical which sets whether cell dots are plotted on top of, versus behind, the violin plots.}

\item{colors}{String vector that sets the colors.
Order of colors should be: `this label`, `this label - pruned`, `any label`.}
}
\value{
a \link[ggplot2]{ggplot} object showing SingleR scores in a dot and/or violin plot representation.
}
\description{
Plot score distributions of labels.
}
\details{
This function creates jitter and violin plots showing the representations of the scores of all cells across a single label or multiple labels,
and may be useful for visualizing and tuning the \code{nmads}, \code{min.diff.med}, and \code{min.diff.next} cutoffs of the \code{\link{pruneScores}} function.
 
The \code{show} input determines what representation to show.  Options are:
\itemize{
\item "scores" = shows the raw scores
\item "delta.med" = shows the difference between max score of a cell and the median of all scores for that cell
\item "delta.next" = shows the difference between best and second-best tuning scores of each cell
}

For a given label X, cells distributions in several categories are shown:
\itemize{
\item Was assigned to label X, and the label was not pruned away.
\item Was assigned to label X, and the label was pruned away.
\item For \code{show="scores"} only, Was assigned as any label, including label X.
}
Each category is grouped and colored separately.
When \code{show = "delta.med"} or \code{show = "delta.next"}, representations are only shown for a cell's final label.

Note that when \code{show="scores"} or \code{show="delta.med"}, the function focuses on initial scores only, i.e., prior to fine tuning.
However, the labels may be defined after fine-tuning in \code{\link{SingleR}} or \code{\link{classifySingleR}}.
Thus, the best score for an individual cell may not be its final label.
}
\examples{
example(SingleR, echo=FALSE)

# To show the distribution of scores grouped by label:
plotScoreDistribution(results = pred)
# We can display a particular label using the label
plotScoreDistribution(results = pred, labels = "B")

# To show the distribution of deltas between cells' maximum and median scores,
#   grouped by label, change `show` to "delta.med":
#   This is useful for checking/adjusting nmads and min.diff.med
plotScoreDistribution(results = pred, show = "delta.med")
# The nmads cutoff can be displayed using show.nmads.cutoff.
plotScoreDistribution(results = pred, show = "delta.med",
    show.nmads.cutoff = 3)

# To show the distribution of deltas between cells' top 2 fine-tuning scores,
#   grouped by label, change `show` to "delta.next":
#   This is useful for checking/adjusting min.diff.next
plotScoreDistribution(results = pred, show = "delta.next")

}
\seealso{
\code{\link{SingleR}}, to generate scores.

\code{\link{pruneScores}}, to remove low-quality labels based on the scores.
}
\author{
Daniel Bunis
}
