% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizations.R
\name{plotScoreHeatmap}
\alias{plotScoreHeatmap}
\title{Plot a score heatmap}
\usage{
plotScoreHeatmap(results, cells.use = NULL, labels.use = NULL,
  clusters = NULL, max.labels = 40, normalize = TRUE,
  cells.order = NULL, order.by.clusters = FALSE, fontsize.row = 9,
  ...)
}
\arguments{
\item{results}{A \linkS4class{DataFrame} containing the output from \code{\link{SingleR}} or \code{\link{classifySingleR}}.}

\item{cells.use}{Integer or character vector specifying the single cells to show.
If \code{NULL}, all cells are presented.}

\item{labels.use}{Character vector indicating what labels to show.
If \code{NULL}, all labels available in \code{results} are presented.}

\item{clusters}{Character vector or factor containing cell cluster assignments, to be shown as annotation in the heatmap.}

\item{max.labels}{Integer scalar specifying the maximum number of labels to show.}

\item{normalize}{Logical specifying whether correlations should be normalized to lie in [0, 1].}

\item{cells.order}{Integer vector specifying the ordering of cells/columns of the heatmap. 
If set, turns off clustering of columns based on scoring.}

\item{order.by.clusters}{Logical scalar specifying if cells should be ordered by \code{clusters} and not by scores.
If set, this takes precedence over \code{cells.order} input.}

\item{...}{Additional parameters for heatmap control passed to \code{pheatmap()}}

\item{silent}{Logical scalar that specifying whether the plot drawn.}
}
\value{
A heatmap of assignment scores is generated on the current graphics device using \pkg{pheatmap}.
}
\description{
Create a heatmap of the \code{\link{SingleR}} assignment scores across all cell-label combinations.
}
\details{
This function creates a heatmap containing the \code{\link{SingleR}} assignment scores for each cell (columns) to each reference label (rows).
Users can then easily identify the high-scoring labels associated with each cell and/or cluster of cells.

If \code{max.labels} is less than the total number of unique labels, only the labels with the largest maximum scores in \code{results} are shown in the plot.
Specifically, the set of scores for each cell is centred and scaled, and the maximum transformed score for each label is used to choose the labels to retain.

If \code{normalize=TRUE}, scores will be linearly adjusted for each cell so that the smallest score is 0 and the largest score is 1.
This is followed by cubing of the adjusted scores to improve dynamic range near 1.
Note that this transformation is done \emph{after} the choice of the top \code{max.labels} labels.
}
\examples{
# Running the SingleR() example.
example(SingleR, echo=FALSE)

# Creating the heatmap.
plotScoreHeatmap(pred)

# Creating a heatmap with clusters.
plotScoreHeatmap(pred, clusters=test$label)

}
\author{
Daniel Bunis, based on code by Dvir Aran.
}
