% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotScoresByCellOrLabel.R
\name{plotScoresByCellOrLabel}
\alias{plotScoresByCellOrLabel}
\alias{plotScoresSingleCell}
\alias{plotScoresSingleLabel}
\alias{plotScoresMultiLabels}
\title{Plot SingleR scoring on a per label or per cell basis.}
\usage{
plotScoresSingleCell(results, cell.id,
  labels.use = levels(as.factor(results$labels)), size = 2,
  colors = c("#F0E442", "#56B4E9", "gray70"))

plotScoresSingleLabel(results, label, size = 0.5, dots.on.top = FALSE,
  colors = c("#F0E442", "#56B4E9", "gray70"))

plotScoresMultiLabels(results, size = 0.2, dots.on.top = FALSE,
  labels.use = levels(as.factor(results$labels)), ncol = 5,
  colors = c("#F0E442", "#56B4E9", "gray70"))
}
\arguments{
\item{results}{A \linkS4class{DataFrame} containing the output from \code{\link{SingleR}} or \code{\link{classifySingleR}}.}

\item{cell.id}{Integer specifying which cell to show for \code{plotScoresSingleCell}}

\item{labels.use}{String vector indicating what labels to show in \code{plotScoresSingleCell} and \code{plotScoresMultiLabel}
If \code{labels.use} is left \code{NULL}, all labels available in \code{results} are presented.}

\item{size}{Scalar, the size of the dots}

\item{colors}{String vector that sets the colors.
Order of colors should be: `this label`, `this label - pruned`, `other label`/`any label`.
Name differently to update the legend.}

\item{label}{String indicating which individual label to plot in \code{plotScoresSingleLabel}
This input will be unnecessary once \code{\link{pruneScores}}'s output is added to the results DataFrame}

\item{dots.on.top}{Logical which sets whether cell dots are plotted on top of, versus behind, the violin plots in \code{plotScoresSingleLabel} and \code{plotScoresMultiLabel}}

\item{ncol}{Integer number of labels to display per row}
}
\value{
Each function returns a \link{ggplot} object showing SingleR scores in a dot and/or violin plot representation.
}
\description{
Plot SingleR scoring on a per label or per cell basis.
}
\details{
NOTE: these functions show initial scores only.
Aside from use of post-fine-tuning labels, they are completely independent of the fine-tuning step.

The \code{plotScoresSingleCell} function creates a dot plot showing the scores of a single cell accross many labels,
with a dotted line added to indicate the median score accross all labels.
It can be used to assess how an individual cell scored versus all/many labels of the reference dataset,
and may be useful for visualizing and tuning the \code{min.diff.med} per-cell cutoff of the \code{\link{pruneScores}} function.

The \code{plotScoresSingleLabel} and \code{plotScoresMultiLabel} functions creates jitter and violin plots showing
the scores of all cells accross either a single label or multiple labels, respectively.
These functions can be used to assess the distribution of scores of all cells for individual labels,
and may be useful for visualizing and tuning the \code{nmads} per-label cutoff of the \code{\link{pruneScores}} function.
All scores for a label are shown in the Cell Calls = "any label".
Scores where this label was the final call for a cell are shown again, separately.
If SingleR calls in \code{results} had been scored for pruning, scores for each label are also separated and colored based on whether cells' calls
were pruned versus not.
}
\section{Functions}{
\itemize{
\item \code{plotScoresSingleCell}: Plot scores accross labels of an individual cells

\item \code{plotScoresSingleLabel}: Plot scores accross labels of an individual cells

\item \code{plotScoresMultiLabels}: Plot scores accross labels of an individual cells
}}

\examples{
example(SingleR, echo=FALSE)

plotScoresSingleCell(results = pred, cell.id = 1)
plotScoresSingleLabel(results = pred, label = "B",
    dots.on.top = TRUE)
plotScoresMultiLabels(results = pred,
    dots.on.top = TRUE, size = 0.5)

}
\seealso{
\code{\link{SingleR}}, to generate scores.

\code{\link{pruneScores}}, to remove low-quality labels based on the scores.
}
\author{
Daniel Bunis
}
