% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{plotScoresByCellOrLabel}
\alias{plotScoresByCellOrLabel}
\alias{plotScoresSingleCell}
\alias{plotScoresSingleLabel}
\alias{plotScoresMultiLabels}
\title{Plot SingleR scoring on a per label or per cell basis.}
\usage{
plotScoresSingleCell(results, cell.id, prune.calls = NULL,
  labels.use = levels(as.factor(results$labels)), size = 2,
  colors = c("#F0E442", "#56B4E9", "gray70", "gray40"))

plotScoresSingleLabel(results, prune.calls = NULL, label, size = 0.5,
  dots.on.top = FALSE, df = NULL, colors = c("#F0E442", "#56B4E9",
  "gray70", "gray40"))

plotScoresMultiLabels(results, prune.calls = NULL, size = 0.2,
  dots.on.top = FALSE, labels.use = levels(as.factor(results$labels)),
  ncol = 5, colors = c("#F0E442", "#56B4E9", "gray70", "gray40"), ...)
}
\arguments{
\item{results}{A \linkS4class{DataFrame} containing the output from \code{\link{SingleR}} or \code{\link{classifySingleR}}.}

\item{cell.id}{Integer specifying which cell to show for \code{plotScoresSingleCell}}

\item{prune.calls}{Logical vector, the output of \code{\link{pruneScores}}.
This input will be unnecessary once \code{\link{pruneScores}}'s output is added to the SingleR.results dataframe}

\item{labels.use}{String vector indicating what labels to show in \code{plotScoresSingleCell} and \code{plotScoresMultiLabel}
If \code{labels.use} is left \code{NULL}, all labels available in \code{results} are presented.}

\item{size}{the size of the dots}

\item{colors}{String vector that sets the colors.  Order of colors should be: `this label`, `this label - pruned`, `other label`, `other label - pruned`.}

\item{label}{String indicating which individual label to plot in \code{plotScoresSingleLabel}}

\item{dots.on.top}{Logical which sets whether cell dots are plotted on top of, versus behind, the violin plots in \code{plotScoresSingleLabel} and \code{plotScoresMultiLabel}}
}
\value{
Dot and violin plot of SingleR scores separated by whether this was the final call or not.
If prune.calls are provided, will also separate and color based on whether cells' final calls were pruned versus not.
}
\description{
Plot SingleR scoring on a per label or per cell basis.
}
\details{
The \code{plotScoresSingleCell} function creates a jitter plot showing the scores of a single cell accross many labels.

The \code{plotScoresSingleLabel} function creates a jitter and violin plot showing the scores of all cells accross a single label.

The \code{plotScoresMultiLabel} function creates \code{plotScoresSingleLabel} plots for multiple individual labels.
}
\section{Functions}{
\itemize{
\item \code{plotScoresSingleCell}: Plot scores accross labels of an individual cells

\item \code{plotScoresSingleLabel}: Plot scores accross labels of an individual cells

\item \code{plotScoresMultiLabels}: Plot scores accross labels of an individual cells
}}

\examples{
example(SingleR, echo=FALSE)
prune <- pruneScores(pred$scores)

plotScoresSingleCell(results = pred, prune.calls = prune, cell.id = 1)
plotScoresSingleLabel(results = pred, prune.calls = prune, label = "B",
    dots.on.top = TRUE)
plotScoresMultiLabels(results = pred,
    dots.on.top = TRUE, size = 0.5)

NULL
}
\author{
Daniel Bunis
}
