% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleR.R
\name{SingleR.CreateObject}
\alias{SingleR.CreateObject}
\title{Wrapper function to create a SingleR object}
\usage{
SingleR.CreateObject(sc.data, ref, clusters, do.main.types = T,
  species = "Human", citation = "-", technology = "-")
}
\arguments{
\item{technology}{}
}
\description{
Wrapper function to create a SingleR object
}
