% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleR.R, R/SingleR.References.R
\docType{package}
\name{SingleR}
\alias{SingleR}
\alias{SingleR-package}
\title{The main SingleR function}
\usage{
SingleR(method = "single", sc_data, ref_data, types, clusters = NULL,
  genes = "de", quantile.use = 0.8, p.threshold = 0.05,
  fine.tune = TRUE, fine.tune.thres = 0.05, sd.thres = 1,
  do.pvals = T, numCores = SingleR.numCores)
}
\arguments{
\item{method}{annotating each single-cell or as a group by cluster: (\code{"single"} or \code{"cluster"})}

\item{sc_data}{the single-cell RNA-seq data set as a matrix with genes as rownames. If the data if from a full-length platform, counts must be normalized to gene length (TPM, RPKM, FPKM, etc.).}

\item{ref_data}{the reference dataset with genes as rownames. Gene names must be in the same format as the single cell data (if sc_data uses genes symbols, ref_data must have the same)}

\item{types}{a list of cell type names corresponding to ref_data. Number of elements in types must be equal to number of columns in ref_data}

\item{clusters}{only if using the "cluster" method. Please provide grouping variables as a factor. The number of elements of clusters must be equal to the number of columns in sc_data}

\item{genes}{list of genes to use for the annotations, or a method for extracting the genes from the data. Available methods are standard deviation and gene dispersion (\code{"sd"} or \code{"de"}). default is "de".}

\item{quantile.use}{correlation coefficients are aggregated for multiple cell types in the reference data set. This parameter allows to choose how to sort the cell types scores, by median (0.5) or any other number between 0 and 1. The default is 0.9.}

\item{p.threshold}{Chi-square outlier detection is used to assess the significance power of the top correlation. Single-cell with an annotation of p-value > p.threshold are designated as "X". Only applies for non fine-tuned annotations.}

\item{fine.tune}{perform the fine tuning step? default is TRUE.}

\item{fine.tune.thres}{the fine tuning step performs the scoring procedure for the top scoring labels using only genes that vary between those cell types in the reference data set. The top labels are those with a score lower than the top score by less than fine.tune.thres}

\item{sd.thres}{if genes=='sd' then this is the threshold for defining a variable gene.}

\item{do.pvals}{compute chi-squared outlier test p-values}

\item{numCores}{Number of cores to use.}
}
\value{
a list with the labels and scores
}
\description{
Given single-cell RNAseq data and reference dataset the function returns the best annotation for each single-cell.
}
