% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleR.R
\name{SingleR.Subset}
\alias{SingleR.Subset}
\title{Subseting a SingleR object}
\usage{
SingleR.Subset(singler, subsetdata)
}
\arguments{
\item{singler}{as SingleR object}

\item{subsetdata}{}
}
\description{
Subseting a SingleR object
}
